updating .sh scripts

+++ reset_data.sh
@@ -9,9 +9,15 @@ fi
 git checkout master
 git pull origin master
 
-rm -r data/daily_csv/
-rm -r data/lifetime_csv/
-rm data/fms.db
+if [ -f data/daily_csv/]; then
+  rm -r data/daily_csv/
+fi
+if [ -f data/lifetime_csv/]; then
+  rm -r data/lifetime_csv/
+fi
+if [ -f data/fms.db]; then
+  rm -r data/fms.db
+fi
 
 cd code
 ./download_and_parse_fms_fixies.py
